+++ changelog.md
@@ -4,6 +4,7 @@
     - Quickly swap between mod lists.
     - Different mod versions can be assigned to different profiles.
   - "Delete Mod" context menu item.
+  - "Edit mod_info.json" context menu item.
   - "Open Folder" menu item now shows all versions installed.
 - Fixed
   - Chipper log viewer not correctly condensing consecutive lines.